Add DriverValues for keys 112-116.
Since keyboard mappings are internally defined in terms of
DriverValues, it's not possible to remap a key unless it has a
DriverValue assigned to it in keys.json.  Moreover, these values need
to be distinct in order for different keys to be remapped
independently.

For the five non-standard keys on the X-Bows layout, each of them uses
the same default driver value as another key, so for these, we need to
define virtual DriverValues to represent them.

This allows the special key names (MBackspace, MEnter, MCtrl, MShift,
RSpace) to be used on the "left-hand side" in a mapping file.  (If you
try to use these names on the "right-hand side", it won't work.)  For
example, the following would swap the middle control and shift keys:

[Base,Layer1,Layer2,Layer3,FnLayer1,FnLayer2,FnLayer3]
MCtrl:LShift
MShift:RCtrl