Rough update for README
Clarified the GUI part and running on Linux – changed sudo for *super user*, as other ways to be root work here (`su`, `doas`, etc.). Also added link to #4 so hopefully it's clearer how to compile it.

It's probably worth rewriting somewhere down the line instead of relying to link to issue, but for now I hope it clears up some questions. :)